requirements.txt
pyzmq
numpy
pyqtgraph
PyQt6
numba
pyinstaller